job_message_writer/backend/requirements.txt
fastapi
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic
alembic==1.11.1
boto3==1.28.1
